+++ src/components/organisms/AlbumList/AlbumList.tsx
@@ -1,15 +1,16 @@
 import React from "react";
 import PropTypes from "prop-types";
+import key from 'weak-key'
 
 import { AlbumCard } from '../../atoms'
 import { ContentList } from '../../molecules'
 
 const AlbumList = ({ songs, audioControl }) => {
   return (
-    <ContentList>
+    <ContentList columns='6'>
       {
         songs.map((song) => 
-          <AlbumCard song={song} audioControl={audioControl} />
+          <AlbumCard song={song} audioControl={audioControl} key={key(song)} />
         )
       }
     </ContentList>