+++ src/theme/index.ts
@@ -1,0 +1,5 @@
+import { theme as SpotifyDark } from './spotifyDark'
+
+export {
+  SpotifyDark
+}