+++ CHANGELOG.md
@@ -6,6 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+## [2.1.2] - 2020.xx.xx
+## Added
+
+
+## [2.1.1] - 2020.10.09
+## Added
+- Add style baseline
+- Change to CSS Grid layout
+- Add SpotifyDark theme
+- Convert App to functional component
+
 ## [2.1.0] - 2020.10.08
 ## Changed
 - Migrate all CSS to CSS-in-JSS
@@ -14,7 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Migrate smaller class components to functional components
 - Add Icon, PlayInfo components
 - @ts-nocheck on all files throwing Typescript errors
-
 
 ## [2.0.2] - 2020.09.29
 ### Changed